# SwaggerDocumentation/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Driving-License-Management-APIs",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Applications/{id}": {
      "get": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationReadDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationReadDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationReadDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationUpdateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationUpdateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestObjectResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestObjectResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestObjectResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationReadDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationReadDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationReadDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Licenses/{id}": {
      "get": {
        "tags": [
          "Licenses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseReadDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseReadDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseReadDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Licenses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LicenseUpdateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LicenseUpdateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LicenseUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseReadDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseReadDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseReadDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Licenses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/LocalDrivingLicenseApplications": {
      "get": {
        "tags": [
          "LocalDrivingLicenseApplications"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocalDrivingLicenseApplicationReadDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocalDrivingLicenseApplicationReadDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocalDrivingLicenseApplicationReadDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "LocalDrivingLicenseApplications"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalDrivingLicenseApplicationUpdateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalDrivingLicenseApplicationUpdateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LocalDrivingLicenseApplicationUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LocalDrivingLicenseApplicationReadDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalDrivingLicenseApplicationReadDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalDrivingLicenseApplicationReadDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/LocalDrivingLicenseApplications/{id}": {
      "get": {
        "tags": [
          "LocalDrivingLicenseApplications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LocalDrivingLicenseApplicationReadDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalDrivingLicenseApplicationReadDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalDrivingLicenseApplicationReadDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "LocalDrivingLicenseApplications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalDrivingLicenseApplicationUpdateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalDrivingLicenseApplicationUpdateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LocalDrivingLicenseApplicationUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LocalDrivingLicenseApplicationReadDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalDrivingLicenseApplicationReadDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalDrivingLicenseApplicationReadDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "LocalDrivingLicenseApplications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplicationReadDTO": {
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "integer",
            "format": "int32"
          },
          "applicantPersonID": {
            "type": "integer",
            "format": "int32"
          },
          "applicantFullName": {
            "type": "string",
            "nullable": true
          },
          "applicationDate": {
            "type": "string",
            "format": "date-time"
          },
          "applicationType": {
            "$ref": "#/components/schemas/ApplicationTypeDTO"
          },
          "applicationStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "applicationStatusName": {
            "type": "string",
            "nullable": true
          },
          "createdByUser": {
            "$ref": "#/components/schemas/UserReadDTO"
          },
          "paidFees": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "ApplicationTypeDTO": {
        "required": [
          "fees",
          "id",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "fees": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "ApplicationUpdateDTO": {
        "required": [
          "applicantPersonID",
          "applicationStatusId",
          "applicationTypeID",
          "creatorUserID",
          "paidFees"
        ],
        "type": "object",
        "properties": {
          "applicationID": {
            "type": "integer",
            "format": "int32"
          },
          "applicantPersonID": {
            "type": "integer",
            "format": "int32"
          },
          "applicationTypeID": {
            "type": "integer",
            "format": "int32"
          },
          "applicationStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "paidFees": {
            "type": "number",
            "format": "float"
          },
          "creatorUserID": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "BadRequestObjectResult": {
        "type": "object",
        "properties": {
          "value": {
            "nullable": true
          },
          "formatters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IOutputFormatter"
            },
            "nullable": true
          },
          "contentTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "declaredType": {
            "$ref": "#/components/schemas/Type"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DriverReadDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "personalInformation": {
            "$ref": "#/components/schemas/PersonDTO"
          },
          "createdBy": {
            "$ref": "#/components/schemas/UserReadDTO"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "IOutputFormatter": {
        "type": "object",
        "additionalProperties": false
      },
...
      "LicenseClassDTO": {
        "required": [
          "classFees",
          "defaultValidityLengthInYears",
          "minimumAllowedAge",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          },
          "minimumAllowedAge": {
            "type": "integer",
            "format": "int32"
          },
          "defaultValidityLengthInYears": {
            "maximum": 255,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "classFees": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "LicenseReadDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "applicationId": {
            "type": "integer",
            "format": "int32"
          },
          "driver": {
            "$ref": "#/components/schemas/DriverReadDTO"
          },
          "licenseClass": {
            "$ref": "#/components/schemas/LicenseClassDTO"
          },
          "issueDate": {
            "type": "string",
            "format": "date"
          },
          "expiryDate": {
            "type": "string",
            "format": "date"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "paidFees": {
            "type": "number",
            "format": "float"
          },
          "issueReason": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "$ref": "#/components/schemas/UserReadDTO"
          }
        },
        "additionalProperties": false
      },
      "LicenseUpdateDTO": {
        "required": [
          "applicationId",
          "createdById",
          "driverId",
          "isActive",
          "issueReason",
          "licenseClassId",
          "notes",
          "paidFees"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "applicationId": {
            "type": "integer",
            "format": "int32"
          },
          "driverId": {
            "type": "integer",
            "format": "int32"
          },
          "licenseClassId": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "minLength": 1,
            "type": "string"
          },
          "paidFees": {
            "type": "number",
            "format": "float"
          },
          "isActive": {
            "type": "boolean"
          },
          "issueReason": {
            "type": "integer",
            "format": "int32"
          },
          "createdById": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "LocalDrivingLicenseApplicationReadDTO": {
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "integer",
            "format": "int32"
          },
          "applicantPersonID": {
            "type": "integer",
            "format": "int32"
          },
          "applicantFullName": {
            "type": "string",
            "nullable": true
          },
          "applicationDate": {
            "type": "string",
            "format": "date-time"
          },
          "applicationType": {
            "$ref": "#/components/schemas/ApplicationTypeDTO"
          },
          "applicationStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "applicationStatusName": {
            "type": "string",
            "nullable": true
          },
          "createdByUser": {
            "$ref": "#/components/schemas/UserReadDTO"
          },
          "paidFees": {
            "type": "number",
            "format": "float"
          },
          "localDrivingLicenseApplicationId": {
            "type": "integer",
            "format": "int32"
          },
          "applicationInfo": {
            "$ref": "#/components/schemas/ApplicationReadDTO"
          },
          "licenseClassInfo": {
            "$ref": "#/components/schemas/LicenseClassDTO"
          }
        },
        "additionalProperties": false
      },
      "LocalDrivingLicenseApplicationUpdateDTO": {
        "required": [
          "applicantPersonID",
          "applicationStatusId",
          "applicationTypeID",
          "creatorUserID",
          "licenseClassId",
          "paidFees"
        ],
        "type": "object",
        "properties": {
          "applicationID": {
            "type": "integer",
            "format": "int32"
          },
          "applicantPersonID": {
            "type": "integer",
            "format": "int32"
          },
          "applicationTypeID": {
            "type": "integer",
            "format": "int32"
          },
          "applicationStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "paidFees": {
            "type": "number",
            "format": "float"
          },
          "creatorUserID": {
            "type": "integer",
            "format": "int32"
          },
          "localDrivingLicenseApplicationId": {
            "type": "integer",
            "format": "int32"
          },
          "licenseClassId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "UserReadDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}